refactoring
-grouped cells's view class
-add view & view controller
-add model's
